Covid Analysis Project Files Uploaded
Covid-19Data.py is the file that scrapes the data off of  'https://covidtracking.com/data/us-daily'. It loops through the html, scraping the listed dates and their corresponding number of cases, and inputs them into the CovidDataFile.csv. Covid Case Predictor.ipynb is the Jupyter Notebook File that takes in the csv and performs the regression analysis on the data, graphing the current data as well as the regression line.